🚚 rename: Rename best-product and most-view-product packages
