+++ src/main/java/com/jokim/sivillage/api/trending/bestproduct/infrastructure/BestProductRepository.java
@@ -1,6 +1,6 @@
-package com.jokim.sivillage.api.trending.product.infrastructure;
+package com.jokim.sivillage.api.trending.bestproduct.infrastructure;
 
-import com.jokim.sivillage.api.trending.product.domain.BestProduct;
+import com.jokim.sivillage.api.trending.bestproduct.domain.BestProduct;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
