+++ src/main/java/com/jokim/sivillage/api/trending/mostviewproduct/application/MostViewProductService.java
@@ -1,0 +1,8 @@
+package com.jokim.sivillage.api.trending.mostviewproduct.application;
+
+import com.jokim.sivillage.api.trending.mostviewproduct.dto.MostViewProductDto;
+import java.util.List;
+
+public interface MostViewProductService {
+    List<MostViewProductDto> getMostViewProduct();
+}
