+++ src/main/java/com/jokim/sivillage/api/trending/mostviewproduct/application/MostViewProductServiceImpl.java
@@ -1,10 +1,10 @@
-package com.jokim.sivillage.api.trending.product.application;
+package com.jokim.sivillage.api.trending.mostviewproduct.application;
 
-import com.jokim.sivillage.api.trending.product.domain.BestProduct;
-import com.jokim.sivillage.api.trending.product.dto.BestProductResponseDto;
-import com.jokim.sivillage.api.trending.product.infrastructure.BestProductRepository;
 import com.jokim.sivillage.api.brand.domain.Brand;
 import com.jokim.sivillage.api.brand.infrastructure.BrandRepository;
+import com.jokim.sivillage.api.trending.mostviewproduct.domain.MostViewProduct;
+import com.jokim.sivillage.api.trending.mostviewproduct.dto.MostViewProductDto;
+import com.jokim.sivillage.api.trending.mostviewproduct.infrastructure.MostViewProductRepository;
 import com.jokim.sivillage.api.product.domain.Product;
 import com.jokim.sivillage.api.product.infrastructure.ProductRepository;
 import com.jokim.sivillage.common.entity.BaseResponseStatus;
@@ -20,26 +20,26 @@ import org.springframework.stereotype.Service;
 @ToString
 @Service
 @RequiredArgsConstructor
-public class BestProductServiceImpl implements BestProductService{
+public class MostViewProductServiceImpl implements MostViewProductService {
 
-    private final BestProductRepository bestProductRepository;
+    private final MostViewProductRepository mostViewProductRepository;
     private final ProductRepository productRepository;
     private final BrandRepository brandRepository;
     @Override
-    public List<BestProductResponseDto> getBestProduct() {
-        List<BestProduct> bestProducts = bestProductRepository.findAll();
+    public List<MostViewProductDto> getMostViewProduct() {
+        List<MostViewProduct> mostViewProducts = mostViewProductRepository.findAll();
         // rankValue 기준으로 오름차순 정렬
-        return bestProducts.stream()
-            .sorted(Comparator.comparing(BestProduct::getRankValue)) // rankValue 기준 오름차순 정렬
-            .map(bestProduct -> {
-                Product product =productRepository.findByProductCode(bestProduct.getProductCode()).
+        return mostViewProducts.stream()
+            .sorted(Comparator.comparing(MostViewProduct::getRankValue)) // rankValue 기준 오름차순 정렬
+            .map(mostViewProduct -> {
+                Product product =productRepository.findByProductCode(mostViewProduct.getProductCode()).
                     orElseThrow(()->new BaseException(BaseResponseStatus.NO_EXIST_PRODUCT));
 
                 Integer discountRate = (int) Math.round(((product.getStandardPrice() - product.getDiscountPrice()) / (double) product.getStandardPrice())*100);
 
                 Brand brand = brandRepository.findByBrandCode(product.getBrandCode()).orElseThrow(()
                     -> new BaseException(BaseResponseStatus.NOT_EXIST_BRAND));
-                return BestProductResponseDto.toDto(bestProduct, product, brand, discountRate);
+                return MostViewProductDto.toDto(mostViewProduct, product, brand, discountRate);
             }) // DTO로 변환
             .toList();
     }
