+++ src/main/java/com/jokim/sivillage/api/trending/mostviewproduct/infrastructure/MostViewProductRepository.java
@@ -1,6 +1,6 @@
-package com.jokim.sivillage.api.trending.mostViewProduct.infrastructure;
+package com.jokim.sivillage.api.trending.mostviewproduct.infrastructure;
 
-import com.jokim.sivillage.api.trending.mostViewProduct.domain.MostViewProduct;
+import com.jokim.sivillage.api.trending.mostviewproduct.domain.MostViewProduct;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
